pare down test for unknown tool events
